+++ Runtime/Scripts/DialogNodeData.cs
@@ -4,11 +4,15 @@ using UnityEngine;
 namespace CheapDialogSystem.Editor.Assets
 {
     [Serializable]
-    public class DialogNodeData
+    public struct DialogNodeData
     {
+        // --- Editor values --- //
         public bool EntryPoint;
         public string NodeGUID;
-        public string DialogueText;
         public Vector2 Position;
+        
+        // --- Runtime values --- //
+        public string DialogTitle;
+        public string DialogText;
     }
 }