+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 17/09/23 - 0.0.2 Update
+
+- Overall code improvement for saving / loading
+- Double click on asset to open window with asset loaded
+- Better multiline edition in content and choice inputs
+- Dialog node creation at center of view
+
 # 10/09/22 - Package creation
 
 - Package creation with GNU GPLv3 license